Add more test coverage
